Remove package ESLint suppressions (Fixes #2089)
