+++ eslint.config.js
@@ -159,14 +159,20 @@ const legacyDirectiveCleanupScopes = [
   'packages/agents/src/**/*.{ts,tsx}', // #2085/#2090
   'packages/cli/src/**/*.{ts,tsx}', // #2086/#2087/#2091
   'packages/tools/src/**/*.{ts,tsx}', // #2088
-  'packages/mcp/src/**/*.{ts,tsx}', // #2089/#2092
-  'packages/auth/src/**/*.{ts,tsx}', // #2089
-  'packages/settings/src/**/*.{ts,tsx}', // #2089
-  'packages/telemetry/src/**/*.{ts,tsx}', // #2089
-  'packages/ide-integration/src/**/*.{ts,tsx}', // #2089
-  'packages/a2a-server/src/**/*.{ts,tsx}', // #2089
-  'packages/policy/src/**/*.{ts,tsx}', // #2089
+  'packages/policy/src/**/*.{ts,tsx}', // #2089 not yet decomposed
   'packages/storage/src/**/*.{ts,tsx}', // #2092
+  // #2089 scope: the six target packages (mcp/auth/settings/telemetry/
+  // ide-integration/a2a-server) still contain other files with legacy
+  // eslint-disable directives. Those packages are kept in legacy scope so
+  // existing directives do not break lint. The target files and extracted
+  // modules are locked in completedDirectiveCleanupScopes below, which
+  // overrides this block for those specific files.
+  'packages/mcp/src/**/*.{ts,tsx}', // #2089/#2092 (non-target files)
+  'packages/auth/src/**/*.{ts,tsx}', // #2089 (non-target files)
+  'packages/settings/src/**/*.{ts,tsx}', // #2089 (non-target files)
+  'packages/telemetry/src/**/*.{ts,tsx}', // #2089 (non-target files)
+  'packages/ide-integration/src/**/*.{ts,tsx}', // #2089 (non-target files)
+  'packages/a2a-server/src/**/*.{ts,tsx}', // #2089 (non-target files)
 ];
 
 const completedDirectiveCleanupScopes = [
@@ -222,6 +228,32 @@ const completedDirectiveCleanupScopes = [
   'packages/core/src/prompt-config/prompt-resolver.ts', // #2082
   'packages/core/src/prompt-config/resolver/**/*.{ts,tsx}', // #2082
   'packages/core/src/runtime/runtimeStateFactory.ts', // #2082
+  // #2089 scope — six target files and their extracted modules are fully
+  // compliant: zero inline eslint-disable directives. Locked to error so
+  // any new directive fails immediately.
+  'packages/a2a-server/src/agent/task.ts', // #2089
+  'packages/a2a-server/src/agent/task-runtime-helpers.ts', // #2089
+  'packages/a2a-server/src/agent/task-support.ts', // #2089
+  'packages/auth/src/oauth-errors.ts', // #2089
+  'packages/ide-integration/src/lsp/lsp-service-client.ts', // #2089
+  'packages/ide-integration/src/lsp/lsp-entry-resolver.ts', // #2089
+  'packages/ide-integration/src/lsp/lsp-status-normalizer.ts', // #2089
+  'packages/mcp/src/client/mcp-client.ts', // #2089
+  'packages/mcp/src/client/mcp-status.ts', // #2089
+  'packages/mcp/src/client/mcp-transport.ts', // #2089
+  'packages/mcp/src/client/mcp-discovery.ts', // #2089
+  'packages/mcp/src/client/mcp-connection.ts', // #2089
+  'packages/mcp/src/client/mcp-oauth-helpers.ts', // #2089
+  'packages/mcp/src/client/mcp-schema-validator.ts', // #2089
+  'packages/mcp/src/client/mcp-callable-tool.ts', // #2089
+  'packages/mcp/src/client/mcp-discovery-helpers.ts', // #2089
+  'packages/settings/src/settings/settingsRegistry.ts', // #2089
+  'packages/settings/src/settings/registry/registry-types.ts', // #2089
+  'packages/settings/src/settings/registry/registry-entries-1.ts', // #2089
+  'packages/settings/src/settings/registry/registry-entries-2.ts', // #2089
+  'packages/settings/src/settings/registry/registry-entries-3.ts', // #2089
+  'packages/telemetry/src/telemetry/types.ts', // #2089
+  'packages/telemetry/src/telemetry/events/*.ts', // #2089
 ];
 
 export default tseslint.config(
@@ -847,50 +879,6 @@ export default tseslint.config(
   },
   // ============================================================================
   // End Issue #1569 C5A
-  // ============================================================================
-  // ============================================================================
-  // Issue #1569: Batch C5B - complexity enforcement
-  // ============================================================================
-  // Promote this rule from warn to error for the specific batch scope.
-  {
-    files: ['packages/a2a-server/src/agent/task.ts'],
-    rules: {
-      complexity: ['error', 25],
-    },
-  },
-  // ============================================================================
-  // End Issue #1569 C5B
-  // ============================================================================
-  // ============================================================================
-  // Issue #1569: Batch C5C - sonarjs/cognitive-complexity enforcement
-  // ============================================================================
-  // Promote this rule from warn to error for the specific batch scope.
-  {
-    files: ['packages/a2a-server/src/agent/task.ts'],
-    rules: {
-      'sonarjs/cognitive-complexity': ['error', 30],
-    },
-  },
-  // ============================================================================
-  // End Issue #1569 C5C
-  // ============================================================================
-  // ============================================================================
-  // Issue #1569: Batch C5D - max-lines enforcement
-  // ============================================================================
-  // Promote this rule from warn to error for the specific batch scope.
-  {
-    files: ['packages/a2a-server/src/agent/task.ts'],
-    rules: {
-      'max-lines': [
-        'error',
-        { max: 800, skipBlankLines: true, skipComments: true },
-      ],
-    },
-  },
-  // ============================================================================
-  // ============================================================================
-  // End Issue #1569 S6A
-  // ============================================================================
   // Issue #1569: Batch S6B - sonarjs/no-ignored-exceptions enforcement
   // ============================================================================
   // Ensure catch blocks handle errors rather than silently ignoring them.
