+++ packages/a2a-server/src/config/config.ts
@@ -263,8 +263,12 @@ export function loadEnvironment(): void {
 
 function findEnvFile(startDir: string): string | null {
   let currentDir = path.resolve(startDir);
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- Classic infinite loop pattern: returns when found, breaks when parentDir === currentDir
-  while (true) {
+  let parentDir = path.resolve(startDir);
+  // Use do/while so the root directory is still probed before exiting,
+  // matching the original while(true) traversal that checked currentDir
+  // before testing whether parentDir === currentDir.
+  do {
+    currentDir = parentDir;
     // prefer gemini-specific .env under GEMINI_DIR
     const geminiEnvPath = path.join(currentDir, GEMINI_CONFIG_DIR, '.env');
     if (fs.existsSync(geminiEnvPath)) {
@@ -274,23 +278,16 @@ function findEnvFile(startDir: string): string | null {
     if (fs.existsSync(envPath)) {
       return envPath;
     }
-    const parentDir = path.dirname(currentDir);
-    if (parentDir === currentDir || !parentDir) {
-      // check .env under home as fallback, again preferring gemini-specific .env
-      const homeGeminiEnvPath = path.join(
-        process.cwd(),
-        GEMINI_CONFIG_DIR,
-        '.env',
-      );
-      if (fs.existsSync(homeGeminiEnvPath)) {
-        return homeGeminiEnvPath;
-      }
-      const homeEnvPath = path.join(homedir(), '.env');
-      if (fs.existsSync(homeEnvPath)) {
-        return homeEnvPath;
-      }
-      return null;
-    }
-    currentDir = parentDir;
+    parentDir = path.dirname(currentDir);
+  } while (parentDir !== currentDir && parentDir !== '');
+  // check .env under home as fallback, again preferring gemini-specific .env
+  const homeGeminiEnvPath = path.join(process.cwd(), GEMINI_CONFIG_DIR, '.env');
+  if (fs.existsSync(homeGeminiEnvPath)) {
+    return homeGeminiEnvPath;
   }
+  const homeEnvPath = path.join(homedir(), '.env');
+  if (fs.existsSync(homeEnvPath)) {
+    return homeEnvPath;
+  }
+  return null;
 }
