+++ packages/ide-integration/src/ide/detect-ide.ts
@@ -24,8 +24,9 @@ export interface IdeInfo {
 }
 
 export function isCloudShell(): boolean {
-  // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing -- intentional falsy coalescing: env vars may be empty string, both cases indicate not in cloud shell
-  return !!(process.env['EDITOR_IN_CLOUD_SHELL'] || process.env['CLOUD_SHELL']);
+  const editor = process.env['EDITOR_IN_CLOUD_SHELL'];
+  const cloudShell = process.env['CLOUD_SHELL'];
+  return Boolean(editor) || Boolean(cloudShell);
 }
 
 export function detectIdeFromEnv(): IdeInfo {
