+++ packages/ide-integration/src/lsp/__tests__/lsp-entry-path.test.ts
@@ -1,8 +1,11 @@
-/* eslint-disable no-console */
 import { existsSync, readFileSync } from 'node:fs';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { describe, expect, it } from 'vitest';
+
+const hasImportMetaResolve =
+  typeof (import.meta as unknown as { resolve?: (s: string) => string })
+    .resolve === 'function';
 
 describe('LSP entry path resolution', () => {
   const moduleDir = dirname(fileURLToPath(import.meta.url));
@@ -23,36 +26,18 @@ describe('LSP entry path resolution', () => {
     expect(existsSync(brokenPath)).toBe(false);
   });
 
-  it('resolves via import.meta.resolve in both source and installed contexts', () => {
-    const resolveImportMeta = (
-      import.meta as unknown as {
-        resolve?: (specifier: string) => string;
-      }
-    ).resolve;
-
-    // eslint-disable-next-line vitest/no-conditional-in-test -- intentional: narrowing/filter/parameterized-test context
-    if (typeof resolveImportMeta === 'function') {
-      let packageUrl: string;
-      try {
-        packageUrl = resolveImportMeta('@vybestack/llxprt-code-lsp');
-      } catch (error) {
-        const err = error as { code?: string };
-        if (
-          err.code === 'MODULE_NOT_FOUND' ||
-          err.code === 'ERR_MODULE_NOT_FOUND'
-        ) {
-          console.warn(
-            'LSP package not installed via npm - this is expected in source tree tests',
-          );
-          return;
+  it.runIf(hasImportMetaResolve)(
+    'resolves via import.meta.resolve when package is installed',
+    () => {
+      const resolveImportMeta = (
+        import.meta as unknown as {
+          resolve: (specifier: string) => string;
         }
-        throw error;
-      }
+      ).resolve;
 
+      const packageUrl = resolveImportMeta('@vybestack/llxprt-code-lsp');
       const packagePath = fileURLToPath(packageUrl);
-      // eslint-disable-next-line vitest/no-conditional-expect -- intentional: narrowing/filter/property-test context
       expect(packagePath).toBeTruthy();
-      // eslint-disable-next-line vitest/no-conditional-expect -- intentional: narrowing/filter/property-test context
       expect(existsSync(packagePath)).toBe(true);
 
       let pkgRoot = dirname(packagePath);
@@ -66,16 +51,7 @@ describe('LSP entry path resolution', () => {
       const srcEntry = join(pkgRoot, 'src', 'main.ts');
       const distEntry = join(pkgRoot, 'dist', 'main.js');
       const entryPath = existsSync(srcEntry) ? srcEntry : distEntry;
-      // eslint-disable-next-line vitest/no-conditional-expect -- intentional: narrowing/filter/property-test context
       expect(existsSync(entryPath)).toBe(true);
-      if (entryPath.endsWith('.ts')) {
-        const content = readFileSync(entryPath, 'utf-8');
-        // eslint-disable-next-line vitest/no-conditional-expect -- intentional: narrowing/filter/property-test context
-        expect(content).toContain('parseBootstrapFromEnv');
-      } else {
-        // eslint-disable-next-line vitest/no-conditional-expect -- intentional: narrowing/filter/property-test context
-        expect(entryPath).toBe(distEntry);
-      }
-    }
-  });
+    },
+  );
 });
